requirements.txt
flask
pytesseract
pillow
google-generativeai